requirements.txt
blinker==1.8.2
boto3==1.34.131
botocore==1.34.131
click==8.1.7
fastavro==1.9.4
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.1
pandavro==1.8.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
s3transfer==0.10.1
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
